[PROJ-118] refactor: pr 리뷰 반영
