+++ src/main/java/com/omnm/hanasset/chat/controller/ChatController.java
@@ -2,8 +2,8 @@ package com.omnm.hanasset.chat.controller;
 
 import com.omnm.hanasset.chat.dto.ChatMessageDTO;
 import com.omnm.hanasset.chat.entity.ChatMessage;
-import com.omnm.hanasset.chat.redis.service.RedisStreamPublisher;
-import com.omnm.hanasset.chat.redis.service.RedisStreamSubscriber;
+import com.omnm.hanasset.chat.redis.RedisStreamPublisher;
+import com.omnm.hanasset.chat.redis.RedisStreamSubscriber;
 import com.omnm.hanasset.chat.utils.ChatMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,12 +30,6 @@ public class ChatController {
     /**
      * 메시지 발행: 채팅방 ID를 기반으로 Stream에 메시지 추가
      */
-//    @PostMapping("/publish")
-//    public String publishMessage(@RequestParam String chatroomId, @RequestBody ChatMessageDTO message) {
-//        publisher.publishMessage(chatroomId, message);
-//        return "Message published to chatroom: " + chatroomId;
-//    }
-
     @MessageMapping("/chat.sendMessage/{roomId}")
     @SendTo("/topic/rooms/{roomId}")
     public ChatMessage sendMessage(@DestinationVariable String roomId, ChatMessage message) {
@@ -45,15 +39,6 @@ public class ChatController {
         // 메시지 처리 로직
         return message;
     }
-
-    /**
-     * 메시지 구독: 채팅방 ID를 기반으로 메시지 소비
-     */
-//    @GetMapping("/consume")
-//    public List<ChatMessageDTO> consumeMessages(@RequestParam String chatroomId) {
-//        List<ChatMessageDTO> consumedMessages = subscriber.consumeMessages(chatroomId);
-//        return consumedMessages; // 메시지 목록 반환
-//    }
 
     /**
      * WebSocket 연결 시 Redis에서 메시지 기록을 불러와 클라이언트로 전송
