+++ src/main/java/com/omnm/hanasset/chat/entity/ChatRoom.java
@@ -47,24 +47,4 @@ public class ChatRoom {
     @Column(name = "reservation_info")
     private String reservationInfo; // JSON 문자열로 저장
 
-    // 예약 정보를 객체로 변환하는 메서드
-    public List<ReservationInfoDTO> getReservationInfoAsList() {
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            return objectMapper.readValue(reservationInfo, new TypeReference<List<ReservationInfoDTO>>() {});
-        } catch (Exception e) {
-            e.printStackTrace();
-            return Collections.emptyList(); // 오류가 발생하면 빈 리스트 반환
-        }
-    }
-
-    // 예약 정보를 객체로 설정하는 메서드
-    public void setReservationInfoAsList(List<ReservationInfoDTO> reservationInfoList) {
-        try {
-            ObjectMapper objectMapper = new ObjectMapper();
-            this.reservationInfo = objectMapper.writeValueAsString(reservationInfoList); // JSON 문자열로 변환
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
 }
