--- src/main/java/com/omnm/hanasset/chat/redis/config/RedisStreamConfig.java
@@ -1,16 +1,0 @@
-package com.omnm.hanasset.chat.redis.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedisStreamConfig {
-
-
-
-}
